+++ src/navigations/Flow1Navigator/index.js
@@ -3,7 +3,7 @@ import { View, Text, Dimensions } from 'react-native'
 import { createStackNavigator } from '@react-navigation/stack'
 import { useNavigation } from '@react-navigation/native'
 import { Ionicons } from '@expo/vector-icons'
-import Flow1 from '../../screens/flow2/Flow2'
+import Flow1 from '../../screens/BenefitListFlow2/Flow2'
 import ReflectionHome from '../../screens/ReflectionHome'
 import GetStarted from '../../screens/BenefitListFlow1/GetStarted'
 import Chat from '../../screens/BenefitListFlow1/Chat'
