+++ src/navigations/Flow2Navigator/index.js
@@ -2,7 +2,7 @@ import React from 'react'
 import { createStackNavigator } from '@react-navigation/stack'
 import { useNavigation } from '@react-navigation/native'
 import { modalOptions } from '../utils'
-import Flow2 from '../../screens/flow2/Flow2'
+import Flow2 from '../../screens/BenefitListFlow2/Flow2'
 import ReflectionHome from '../../screens/ReflectionHome'
 import Header from '../../components/Header'
