+++ src/navigations/ShortTermNavigator/index.js
@@ -2,8 +2,8 @@ import React, { useState } from 'react'
 import { createStackNavigator } from '@react-navigation/stack'
 import { useNavigation } from '@react-navigation/native'
 import { modalOptions } from '../utils'
-import ShortTermBenefits from '../../screens/flow2/ShortTermBenefits'
-import ShortTermBenefitModal from '../../screens/flow2/ShortTermBenefitModal'
+import ShortTermBenefits from '../../screens/BenefitListFlow2/ShortTermBenefits'
+import ShortTermBenefitModal from '../../screens/BenefitListFlow2/ShortTermBenefitModal'
 import ShortTermHeader from '../../components/ShortTermHeader'
 import ShortTermModalHeader from '../../components/ShortTermModalHeader'
